\name{CANSworks-package}
\alias{CANSworks-package}
\alias{CANSworks}
\docType{package}

\title{
Manipulate and analyze CANS data
}

\description{
CANS (Child and Adolescent Strengths and Needs), was developed and copyright by John S. Lyons PhD. 
Massachusetts uses an adaptation of the CANS in a variety of public behavioral health services, 
including those paid under the state Medicaid program, MassHealth. This package provides functions to 
(1) import into R CANS data exported from the state's web-based CANS application, and 
(2) to manipulate, analyze and report from the imported data. 
The package was created by J. Simons and is not required, endorsed, supported or warranteed 
in any way by the Commonwealth of Massachusetts.
}

\details{
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\references{
www.praedfoundtation.org
www.mass.gov/masshealth/cans
}

\examples{
% none at this time
}
