\name{check_CANS_export}
\alias{check_CANS_export}
\alias{check_multiple_CANS_exports}

\title{
Checks field length of CANS export file
}

\description{
Checks field length of CANS export file
}

\usage{
check_CANS_export()
}

\value{
Used for its byproducts. Results of check are reported to console.
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\note{
}

\examples{
% none at this time
}