\name{check_multiple_CANS_exports}
\alias{check_multiple_CANS_exports}

\title{
Checks multiple CANS exports}

\description{
Checks multiple CANS exports
}

\usage{
check_multiple_CANS_exports()
}

\value{
Used for its byproducts. Result is reported to console.
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\examples{
## no examples at this time
}
