\name{copy_to_xl}
\alias{copy_to_xl}

\title{
Copy R object to Excel
}
\description{
Copy R object to clipboard in format that pastes nicely in Excel. 
See the R reference card bu Tom Short
}

\usage{
copy_to_xl(x)
}

\arguments{
  \item{x}{an R object
}
}

\value{
Used for its byproduct, which is to place data on the Windows clipboard.
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\examples{
no examples at present
}
