\name{get_one_CANS_per_person}
\alias{get_one_CANS_per_person}
\title{get one CANS per person}

\usage{
get_one_CANS_per_person(x,
most_recent = TRUE)
}
\arguments{
  \item{x}{
}
  \item{most_recent}{
}
}



\author{Jack Simons <jack.simons@state.ma.us>}





