\name{make_model_formula}
\alias{make_model_formula}
\title{make model formula}

\usage{
make_model_formula(leftvars = 0,
rightvars = 0,
data)
}
\arguments{
  \item{leftvars}{
}
  \item{rightvars}{
}
  \item{data}{
}
}



\author{Jack Simons <jack.simons@state.ma.us>}





