\name{paste_from_xl}
\alias{paste_from_xl}

\title{paste_from_xl

}
\description{Creates an R object from the Windows clipboard. See Tom Short's R reference card
}

\usage{
paste_from_xl()
}

\value{
Returns an R object pasted from an Excel spreadsheet
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\examples{
# none now
}
