\name{rating.summary}
\alias{rating.summary}
\title{rating summary}

\usage{
rating.summary(cans,
obs.id = 1,
person.id = 2,
obs.date = 32,
items = c(54:64,
66:75,
77:88))
}
\arguments{
  \item{cans}{
}
  \item{obs.id}{
}
  \item{person.id}{
}
  \item{obs.date}{
}
  \item{items}{
}
}



\author{Jack Simons <jack.simons@state.ma.us>}





