\name{rating_cts}
\alias{rating_cts}

\title{
count the number of CANS items rated 0,1,2, and 3
}

\description{
count the number of CANS items rated 0,1,2, and 3
}

\usage{
rating_cts(cans, histogram = TRUE)
}

\arguments{
  \item{cans}{
a dataframe with CANS ratings in some columns
}
  \item{histogram}{
logical, if TRUE a histogram is plotted
}
}

\details{
}

\value{
a dataframe of same length as the original dataframe
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\examples{
% none at this time
}