\name{read_cans}
\alias{read_cans}
\title{read cans}

\usage{
read_cans(tool = c("Under5",
"5Over",
""),
status = c("COMPLETE",
"DOCUMENTED ON PAPER",
"INCOMPLETE BUT FINAL",
""),
filename = "choose")
}
\arguments{
  \item{tool}{
}
  \item{status}{
}
  \item{filename}{
}
}



\author{Jack Simons <jack.simons@state.ma.us>}





