\name{replace_factor_with_random_id}
\alias{replace_factor_with_random_id}
\title{replace factor with random id}

\usage{
replace_factor_with_random_id(cans,
target_col)
}
\arguments{
  \item{cans}{
}
  \item{target_col}{
}
}



\author{Jack Simons <jack.simons@state.ma.us>}





