\name{sample_cans}
\alias{sample_cans}

\title{
sample CANS
}

\description{Samples records from a CANS dataframe
}

\usage{
sample_cans(cans, sample.size)
}

\arguments{
  \item{cans}{
cans is the dataframe from which you want to sample
}
  \item{sample.size}{
sample.size is the number of records you want to get
}
}

\details{
}

\value{
returns a dataframe of CANS records sampled from the original dataframe
}

\references{
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\note{
}

\examples{
% none at this time
}