\name{select_items}
\alias{select_items}
\title{
makes new CANS dataframe with only certain columns retained -- deprecated
}
\description{
makes new CANS dataframe with only certain columns retained -- deprecated
}
\usage{
select_items(cans, select.items.csv)
}
\arguments{\item{cans}{
a CANS dataframe
}\item{select.items.csv}{
a csv format file that indicates which columns to retain
}
}
\details{
}
\value{
a CANS dataframe
}
\references{
}
\author{
Jack Simons <jack.simons@state.ma.us>
}
\note{
}
\examples{
% none at this time
}