\name{stamp}
\alias{stamp}

\title{
convenient author contact info and timestamp
}
\description{
shortcut for documenting scripts
}
\usage{
stamp(author = "Jack Simons <jack.simons@state.ma.us>")
}
\arguments{
  \item{author}{
author contact info
}
}
\details{
}
\value{
returns a character string
}
\references{
}
\author{
Jack Simons <jack.simons@state.ma.us>
}
\note{
}
\examples{
% none at this time
}