\name{str}
\alias{str}

\title{
Changes the list length for convenience
}

\description{
Changes the list length for convenience
}

\usage{
str(x)
}

\arguments{
  \item{x}{
An R object
}
}

\value{
Same as returned by the str() function
}

\author{
Jack Simons <jack.simons@state.ma.us>
}

\examples{
# none at this time
}